+++ troml_dev_status/__about__.py
@@ -11,7 +11,7 @@ __all__ = [
 ]
 
 __title__ = "troml-dev-status"
-__version__ = "0.4.1"
+__version__ = "0.4.2"
 __keywords__ = ["development status", "trove classifiers", "python packaging"]
 __description__ = "Objectively infer PyPI Development Status classifiers from code and release artifacts."
 __readme__ = "README.md"